all fixes, except looping textures, testing to see which version using.

+++ G53GRA.Framework/G53GRA.Framework/Framework/Engine/Scene.cpp
@@ -1,7 +1,7 @@
-#include "Animation.h"
-#include "Input.h"
+#include "../Interface/Animation.h"
+#include "../Interface/Input.h"
 #include "Scene.h"
-#include "DisplayableObject.h"
+#include "../Interface/DisplayableObject.h"
 
 Texture Scene::texture = Texture();		// Initialise Texture handler
 Camera Scene::camera = Camera();        // Initialise Camera object
@@ -16,9 +16,9 @@ Scene::Scene(int argc, char** argv, const char *title, const int& windowWidth, c
  */
 Scene::~Scene()
 {
-	for(DisplayableObject* obj : objects)
+	for (DisplayableObject* obj : objects)
 	{
-		if(obj != NULL)
+		if (obj != NULL)
 			delete obj;
 		obj = NULL;
 	}
@@ -84,7 +84,9 @@ void Scene::Reshape(int w, int h)
 
 void Scene::Projection(){
 	// Default orthographic projection
-	glOrtho(-windowWidth / 2, windowWidth / 2, -windowHeight / 2, windowHeight / 2, 1.0, 1000.0);
+	glOrtho(static_cast<GLdouble>(-windowWidth) / static_cast<GLdouble>(2.0), static_cast<GLdouble>(windowWidth) / static_cast<GLdouble>(2.0),
+			static_cast<GLdouble>(-windowHeight) / static_cast<GLdouble>(2.0), static_cast<GLdouble>(windowHeight) / static_cast<GLdouble>(2.0),
+			static_cast<GLdouble>(1.0), static_cast<GLdouble>(1000.0));
 }
 
 void Scene::Update(const double& deltaTime)
@@ -111,11 +113,11 @@ void Scene::HandleKey(unsigned char key, int state, int x, int y)
 	// State is 1 for key press, 0 for key release
 	Input* input_obj;
 	camera.HandleKey(key, state, x, y);
-	for(DisplayableObject* obj : objects)
+	for (DisplayableObject* obj : objects)
 	{
 		input_obj = dynamic_cast<Input*>(obj);
-		if(input_obj != NULL)
-			input_obj -> HandleKey(key, state, x, y);
+		if (input_obj != NULL)
+			input_obj->HandleKey(key, state, x, y);
 	}
 }
